factoring mainview away from app
